feat: add egapro config file
